Create OLED_Reticle_MPU9250_SAMD51.ino
Refactor the code by adding the additional hysteresis component. The triangle reticle should "float" in the X, Y axes range based off the movement of the device as sampled by the MPU. Also suggest a file name upon completion. Reproduce the code in it's entirety rather than a snippet.

Adding the hysteresis behavior can be achieved by implementing a simple low-pass filter on the accelerometer data, which allows the reticle to "float" more naturally. Here is the complete Arduino code:

In this code, the previous acceleration data is stored and used to calculate the new position of the reticle. The ALPHA parameter defines the weighting of the new data versus the old data. A lower value will make the reticle move more slowly and smoothly.