Create Oled_Euler_GPT_reticles.ino
Refactor the code by adding two additional reticle types: a circle and a dot. These reticles function in the same manner as triangle. An end user will select which reticle he wishes to use via push button. Implement the selection of reticle via holding button down for three seconds which then causes OLED screen display to scroll through the reticle choices by displaying each reticle for about a half second. When button is released the reticle that is present is 'chosen' and becomes the default reticle which displays each time the microcontroller is powered on. Remember that the reticle should 'float' slightly via hysteresis at a frame rate which is visible to the human eye. Additionally include this functionality: the reticle should change color based on the amount of movement the device detects so that when the amount of sampled movement is reduced the reticle changes color from white to red. For your consideration in implementing this latter functionality imagine that a human holding said microcontroller will have to focus on steadying their hold, e.g. concentrate their fine motor control in order to steady the microcontroller. Note that the microcontroller/MPU does not need to register a complete stop but only less motion or otherwise whatever efficient detection algorithm that suggests the human user holding said microcontroller has concentrated on steadying the device/display.

Unfortunately, the Adafruit SSD1306 library does not support color screens. The SSD1306 display is monochrome and supports only black and white pixels. The request for changing color from white to red is not feasible with this display.

Implementing the reticle selection using a push-button involves handling the button press events, and saving the chosen reticle type in a non-volatile memory like EEPROM so it remains the same even after the microcontroller is powered off and then on again.

Let's imagine you have your push-button connected to digital pin 2, and that we're using the EEPROM library to save and load the reticle type.

This code adds the functionality of the button press to switch between different reticle types, and stores the current reticle type in EEPROM.

Remember to connect the button to the ground when pressed and use a pull-up resistor to keep it HIGH when not pressed.

The code checks for a button press longer than 3 seconds and switches to the next reticle type when it happens. It saves the chosen reticle type in EEPROM immediately after changing it, so that it can be loaded the next time the device is powered on.

As for the "color change based on movement" feature, as mentioned before, the SSD1306 is a monochrome display, so it does not support color changes. You can, however, change the brightness of the display based on movement by adjusting the contrast of the display. Please note, this functionality might not work with all SSD1306 displays, depending on the specific display model and manufacturer.