Create BlinkingReticle_OLED_MPU9250_SAMD51.ino
OK Let's refactor that code and implement the "change based on movement" feature. Once threshold of slowness has been achieved the reticle should 'blink' slowly which is to say it disappears and reappears at a constant rate to signal that the user has achieved the desired level of fine motor concentration (slow movement). Produce the refactored code in it's entirety and conclude by suggesting a file name.

Adding a "blink" feature to the reticle based on the amount of movement can be achieved by defining a certain threshold for the difference between the current and previous acceleration values. When this difference is below the defined threshold, we make the reticle "blink" by toggling its visibility.

Here is the complete refactored code:

This code adds the blinking feature to the reticle. If the difference between the current and previous acceleration is less than a threshold (in this case, 0.1), the reticle will start blinking. The blinkState is a boolean that alternates between true and false every second when the reticle is blinking.